Update README.md
Assume a kitchen has n chief working. If each chef can claim atmost ‘r’ ingredients(each 
ingredients is kept in a jar),then find out the total number of ingredients that are present in the 
kitchen having deadlock free execution of work. Write a code to implement the above 
situation.
Expected outcomes:
1. Deadlock free scenario: The program should be able to create a deadlock 
free situation as per the above mentioned scenario
2. Explanation about the deadlock free algorithm has been implemented
3. At least 5 test cases should be mentioned to test the scenario
Simulation results: The program should run the simulation for a set of values and and should 
display both deadlock and deadlock free environment, Also there should be valid logic 
behind choosing a set of ingredients for a deadlock free scenario